+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # CHANGELOG
 
-## Unreleased
+## 0.8.5
 
 - [#61](https://github.com/georust/gpx/pull/61): Allow custom xml::EventWriter in write(add `write_with_event_writer`)
 - [#62](https://github.com/georust/gpx/pull/62): Implementing serde (De)Serialize for GPX structs
+- [#42](https://github.com/georust/gpx/pull/62): Track extensions
 
 ## 0.8.4